requirements.txt
beautifulsoup4==4.5.1
dateparser==0.4.0
Django==1.10.1
django-filter==0.15.2
django-filters==0.2.1
djangorestframework==3.4.7
factory-boy==2.7.0
fake-factory==0.7.2
jdatetime==1.8.1
lxml==3.6.4
Markdown==2.6.7
numpy==1.11.2
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.7
regex==2016.9.22
requests==2.11.1
ruamel.yaml==0.12.14
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
umalqurra==0.2
pep8
flake8
